[Embedding] Refactor and separate filter policices into standalone files.
