+++ tensorflow/core/framework/embedding/single_tier_storage.h
@@ -152,7 +152,7 @@ class SingleTierStorage : public Storage<K, V> {
       std::vector<int64>* version_list,
       std::vector<int64>* freq_list,
       const EmbeddingConfig& emb_config,
-      EmbeddingFilter<K, V, EmbeddingVar<K, V>>* filter,
+      FilterPolicy<K, V, EmbeddingVar<K, V>>* filter,
       embedding::Iterator** it) override {
     std::vector<ValuePtr<V>*> value_ptr_list;
     std::vector<K> key_list_tmp;
